Revert "Implemented first pass of A*"
This reverts commit 659b96ff8054770a4996a9c73af51a3713b2e8cf.

+++ source/engine/resources/TextureResource.cpp
@@ -38,7 +38,7 @@ GLuint TextureResource::GetGLTextureId() const
 
 ///------------------------------------------------------------------------------------------------
 
-const glm::ivec2& TextureResource::GetDimensions() const
+const glm::vec2& TextureResource::GetDimensions() const
 {
     return mDimensions;
 }
